requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.20.0
pillow>=9.0.0
requests>=2.28.0
numpy>=1.24.0
opencv-python>=4.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0
